+++ python/python/tests/test_async_client.py
@@ -911,6 +911,50 @@ class TestCommands:
 
     @pytest.mark.parametrize("cluster_mode", [True, False])
     @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_blpop(self, redis_client: TRedisClient):
+        key1 = f"{{test}}-f{get_random_string(10)}"
+        key2 = f"{{test}}-f{get_random_string(10)}"
+        value1 = "value1"
+        value2 = "value2"
+        value_list = [value1, value2]
+
+        assert await redis_client.lpush(key1, value_list) == 2
+        assert await redis_client.blpop([key1, key2], 0.5) == [key1, value2]
+
+        # Redis versions < 7.0.0 do not accept floating point values for the timeout, so will be rounded
+        timeout = 1.0 if check_if_server_version_lt(redis_client, "7.0.0") else 0.001
+        with pytest.raises(TimeoutError):
+            await redis_client.blpop([key2], timeout)
+
+        assert await redis_client.set("foo", "bar")
+        with pytest.raises(RequestError) as e:
+            await redis_client.blpop(["foo"], timeout)
+        assert "Operation against a key holding the wrong kind of value" in str(e)
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
+    async def test_brpop(self, redis_client: TRedisClient):
+        key1 = f"{{test}}-f{get_random_string(10)}"
+        key2 = f"{{test}}-f{get_random_string(10)}"
+        value1 = "value1"
+        value2 = "value2"
+        value_list = [value1, value2]
+
+        assert await redis_client.lpush(key1, value_list) == 2
+        assert await redis_client.brpop([key1, key2], 0.5) == [key1, value1]
+
+        # Redis versions < 7.0.0 do not accept floating point values for the timeout, so will be rounded
+        timeout = 1.0 if check_if_server_version_lt(redis_client, "7.0.0") else 0.001
+        with pytest.raises(TimeoutError):
+            await redis_client.brpop([key2], timeout)
+
+        assert await redis_client.set("foo", "bar")
+        with pytest.raises(RequestError) as e:
+            await redis_client.brpop(["foo"], timeout)
+        assert "Operation against a key holding the wrong kind of value" in str(e)
+
+    @pytest.mark.parametrize("cluster_mode", [True, False])
+    @pytest.mark.parametrize("protocol", [ProtocolVersion.RESP2, ProtocolVersion.RESP3])
     async def test_sadd_srem_smembers_scard(self, redis_client: TRedisClient):
         key = get_random_string(10)
         value_list = ["member1", "member2", "member3", "member4"]